DRTII-1261 Tidying
